+++ package.json
@@ -49,7 +49,6 @@
     "npm-run-all": "^4.1.5",
     "nyc": "^14.0.0",
     "semantic-release": "^15.13.19",
-    "source-map-support": "^0.5.19",
     "trash-cli": "^1.4.0",
     "ts-node": "^8.0.3",
     "typescript": "^3.9.7",
