+++ src/data/cards.ts
@@ -4,6 +4,6 @@ type CardType = {dbfId: number; id: string; name: string};
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const cardsJson: CardType[] = require('../../data/cards.json');
-const cardData = keyBy(cardsJson, c => c.id) as {[key: string]: CardType};
+const cardData = keyBy(cardsJson, c => c.id) as {[key: string]: CardType | undefined};
 
 export default cardData;
