+++ src/index.ts
@@ -1,5 +1,3 @@
-import 'source-map-support/register';
-
 export * from './LogWatcher';
 export {Events} from './line-parsers';
 export {GameState, Player} from './GameState';
